requirements.txt
h5py
joblib
ninja
numba
pandas
pyyaml
scipy
tqdm
visdom
